# user-service/src/main/resources/openapi.yaml
openapi: 3.0.1
info:
  termsOfService: http://swagger.io/terms/
  title: User Service
  version: 0.0.1
servers:
- url: /
...
paths:
...
  /suppliers:
    get:
      description: Returns a list of suppliers
      operationId: listSuppliers
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Supplier'
                type: array
          description: successful operation
      summary: List suppliers
      tags:
      - suppliers
      x-accepts: application/json
      x-tags:
      - tag: suppliers
    post:
      operationId: createSupplier
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SupplierCreate'
        description: Supplier fields
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Supplier'
          description: successful supplier creation
        "400":
          content: {}
          description: Invalid input
      summary: Create a new supplier
      tags:
      - suppliers
      x-codegen-request-body-name: body
      x-content-type: application/json
      x-accepts: application/json
      x-tags:
      - tag: suppliers
  /suppliers/{supplierId}:
    delete:
      description: Deletes a single supplier by id
      operationId: deleteSupplier
      parameters:
      - description: ID of the supplier to delete
        explode: false
        in: path
        name: supplierId
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      responses:
        "200":
          description: successfully deleted the supplier
        "404":
          description: No supplier with the id was found
      summary: Delete supplier by id
      tags:
      - suppliers
      x-accepts: application/json
      x-tags:
      - tag: suppliers
    get:
      description: Gets a single supplier by id
      operationId: getSupplier
      parameters:
      - description: ID of supplier to return
        explode: false
        in: path
        name: supplierId
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Supplier'
          description: successful listing of suppliers
        "404":
          description: supplier with id inputted not found
      summary: Get supplier by id
      tags:
      - suppliers
      x-accepts: application/json
      x-tags:
      - tag: suppliers
    put:
      description: Update a single supplier by id
      operationId: updateSupplier
      parameters:
      - description: ID of the supplier to update
        explode: false
        in: path
        name: supplierId
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SupplierUpdate'
        description: Supplier fields to update
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Supplier'
          description: successfully updated the supplier
        "404":
          description: No supplier with the id was found
      summary: Update supplier by id
      tags:
      - suppliers
      x-accepts: application/json
      x-tags:
      - tag: suppliers
      x-content-type: application/json
components:
  schemas:
...
    SupplierUpdate:
      example:
        name: Ricardo Meat
        product_type: MEAT
      properties:
        name:
          example: Ricardo Meat
          type: string
        product_type:
          example: MEAT
          type: string
      type: object
    SupplierCreate:
      example:
        name: Ricardo Meat
        product_type: MEAT
      properties:
        name:
          type: string
        product_type:
          type: string
      required:
      - name
      - product_type
      type: object
    Supplier:
      example:
        id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        name: Ricardo's Meat
        product_type: MEAT
        created_at: 2022-07-21T02:05:41.527664Z
        updated_at: 2022-07-21T02:05:41.527664Z
      properties:
        id:
          format: uuid
          type: string
        name:
          type: string
        product_type:
          type: string
        created_at:
          format: date-time
          type: string
        updated_at:
          format: date-time
          type: string
      required:
      - created_at
      - id
      - name
      - product_type
      - updated_at
      type: object